+++ tests/Feature/PendingUpdatesTest.php
@@ -28,6 +28,10 @@ class PendingUpdatesTest extends FeatureTestCase
 
     public static function setUpBeforeClass(): void
     {
+        if (! getenv('STRIPE_SECRET')) {
+            return;
+        }
+
         parent::setUpBeforeClass();
 
         static::$productId = self::stripe()->products->create([